pyproject.toml
[project]
name = "dudek"
version = "0.1.0"
description = "SoccerNet challenge development kit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "huggingface-hub>=0.26.2",
    "jsonlines>=4.0.0",
    "kornia>=0.7.4",
    "moviepy>=2.1.1",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
    "soccernet>=0.1.61",
    "tensorboard>=2.18.0",
    "timm>=1.0.12",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "wandb>=0.19.1",
]
[dependency-groups]
dev = [
    "notebook>=7.2.2",
    "black>=24.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["notebooks", "scripts"]


[project.scripts]
bas-frame-extract = "dudek.scripts.extract:cli"
bas-tdeed-train = "dudek.scripts.tdeed:cli"